3.6 Stock Levels by SKU - backend 1.0
I added a utils.py file which contains the check_sku_alerts function. The function pulls from inventory and should dynamically display the parts and their attributes. At the moment, the alert is only generated when manually queried (TO DO: make dynamic alerts based on threshold levels). There is a filter that can be implemented as well (No filter is applied by default.

Defined the Alert model to store alerts when they are generated.

Updated urls.py and views.py